change SPI to LL library
Noticed that there was a lot of latency in the HAL transmit
function. It was taking 35us to send two words, with a 5us delay
in between them. The acutal transmissions take up 7.75us. So that's
22.25us of delay or roughly 5.55us before and after a chip select
line toggle. With the LL drivers, the total transmit time is 17us.
That makes for rougly 1.1us before and after a chip select line
toggle. The use of the LL driver has doubled this SPI transmission
performance.